requirements_base_beta.txt
appdirs==1.4.3
asgi-redis==1.4.3
asgiref==1.1.2
autobahn==17.9.3
Babel==2.3.4
bleach==2.1.2
boto==2.42.0
channels==1.1.8
colorama==0.3.7
contextlib2==0.5.4
coverage==4.2
daphne==1.3.0
decorator==4.1.2
dj-database-url==0.4.1
Django==1.8.8
django-appconf==1.0.2
django-countries==4.0
django-easymoney==0.7.1
django-environ==0.4.2
django-extensions==1.7.8
django-floppyforms==1.7.0
django-idmap==1.0
django-vanilla-views==1.0.4
djangorestframework==3.4.6
entrypoints==0.2.3
honcho==0.7.1
html5lib==1.0.1
huey==1.2.0
IPy==0.83
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.0
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.1
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
mock==2.0.0
msgpack-python==0.4.8
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
olefile==0.44
otree-boto2-shim==0.3.2
otree-save-the-change==1.1.3
packaging==16.8
pandocfilters==1.4.2
parso==0.1.1
pbr==1.10.0
pexpect==4.3.1
pickleshare==0.7.4
Pillow==4.1.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.4.31
Pygments==2.2.0
pyparsing==2.2.0
pytest==2.9.2
pytest-django==3.0.0
python-dateutil==2.6.1
python-redis-lock==3.2.0
pytz==2017.2
pyzmq==16.0.3
qtconsole==4.3.1
raven==5.25.0
redis==2.10.6
requests==2.11.1
schema==0.6.2
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
Twisted==17.9.0
txaio==2.8.2
unicodecsv==0.14.1
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==3.2.1
widgetsnbextension==3.1.0
ws4py==0.3.5
XlsxWriter==0.9.3
zope.interface==4.4.3
